+++ sdk/sdk/api_client/models/aws_glue_config.py
@@ -6,6 +6,8 @@ from typing import Any, TypeVar
 from attrs import define as _attrs_define
 from attrs import field as _attrs_field
 
+from ..types import UNSET, Unset
+
 T = TypeVar("T", bound="AWSGlueConfig")
 
 
@@ -17,12 +19,14 @@ class AWSGlueConfig:
         database_name (str):
         bucket_identifier (str):
         s3_path (str):
+        athena_workgroup_template (str | Unset):  Default: ''.
     """
 
     identifier: str
     database_name: str
     bucket_identifier: str
     s3_path: str
+    athena_workgroup_template: str | Unset = ""
     additional_properties: dict[str, Any] = _attrs_field(init=False, factory=dict)
 
     def to_dict(self) -> dict[str, Any]:
@@ -34,6 +38,8 @@ class AWSGlueConfig:
 
         s3_path = self.s3_path
 
+        athena_workgroup_template = self.athena_workgroup_template
+
         field_dict: dict[str, Any] = {}
         field_dict.update(self.additional_properties)
         field_dict.update(
@@ -44,6 +50,8 @@ class AWSGlueConfig:
                 "s3_path": s3_path,
             }
         )
+        if athena_workgroup_template is not UNSET:
+            field_dict["athena_workgroup_template"] = athena_workgroup_template
 
         return field_dict
 
@@ -58,11 +66,14 @@ class AWSGlueConfig:
 
         s3_path = d.pop("s3_path")
 
+        athena_workgroup_template = d.pop("athena_workgroup_template", UNSET)
+
         aws_glue_config = cls(
             identifier=identifier,
             database_name=database_name,
             bucket_identifier=bucket_identifier,
             s3_path=s3_path,
+            athena_workgroup_template=athena_workgroup_template,
         )
 
         aws_glue_config.additional_properties = d